modified motion control to generate line segments for 16bit bresenham algorithm
-modified motion control to generate line segments for 16bit bresenham algorithm. This splits longer lines into smaller segments that can be computed with 16bit math.
-moved DSS_OVERSAMPLING check to config help
-modified some memcpy and memset size calculations

+++ src/core/planner.h
@@ -44,8 +44,8 @@ extern "C"
         uint32_t line;
 #endif
         uint8_t dirbits;
-        uint32_t steps[STEPPER_COUNT];
-        uint32_t total_steps;
+        uint16_t steps[STEPPER_COUNT];
+        uint16_t total_steps;
         uint8_t step_indexer;
 
         float entry_feed_sqr;
